move accesstoken from cookies into params

+++ src/roles/roles.resolver.ts
@@ -3,16 +3,19 @@ import { User } from 'src/users/user.model';
 import { UsersService } from 'src/users/users.service';
 import { Role } from './role.model';
 import { RolesService } from './roles.service';
-import { forwardRef, Inject, UseGuards } from '@nestjs/common';
-import { CurrentUser, GqlAuthGuard } from 'src/auth/gql-auth.guard';
+import { BadRequestException, forwardRef, Inject } from '@nestjs/common';
 import { PUB_SUB } from 'src/pub-sub/pub-sub.module';
 import { RedisPubSub } from 'graphql-redis-subscriptions';
 import { Arrow } from 'src/arrows/arrow.model';
 import { ArrowsService } from 'src/arrows/arrows.service';
+import { ConfigService } from '@nestjs/config';
+import { JwtService } from '@nestjs/jwt';
 
 @Resolver(() => Role)
 export class RolesResolver {
   constructor(
+    private readonly configService: ConfigService,
+    private readonly jwtService: JwtService,
     private readonly rolesService: RolesService,
     @Inject(forwardRef(() => UsersService))
     private readonly usersService: UsersService,
@@ -35,14 +38,21 @@ export class RolesResolver {
     return this.arrowsService.getArrowById(role.arrowId);
   }
 
-  @UseGuards(GqlAuthGuard)
   @Mutation(() => Role, {name: 'inviteRole'})
   async inviteRole(
-    @CurrentUser() user: User,
+    @Args('accessToken') accessToken: string,
     @Args('sessionId') sessionId: string,
     @Args('userName') userName: string,
     @Args('arrowId') arrowId: string,
   ) {
+    const payload = this.jwtService.verify(accessToken, {
+      secret: this.configService.get('JWT_ACCESS_TOKEN_SECRET')
+    });
+    const user = await this.usersService.getUserById(payload.userId);
+    if (!user) {
+      throw new BadRequestException('Invalid accessToken');
+    }
+
     const role = await this.rolesService.inviteRole(user.id, userName, arrowId);
     const arrow = await this.arrowsService.getArrowById(arrowId);
 
@@ -69,13 +79,20 @@ export class RolesResolver {
     return role;
   }
 
-  @UseGuards(GqlAuthGuard)
   @Mutation(() => Role, {name: 'requestRole'})
   async requestRole(
-    @CurrentUser() user: User,
+    @Args('accessToken') accessToken: string,
     @Args('sessionId') sessionId: string,
     @Args('arrowId') arrowId: string,
   ) {
+    const payload = this.jwtService.verify(accessToken, {
+      secret: this.configService.get('JWT_ACCESS_TOKEN_SECRET')
+    });
+    const user = await this.usersService.getUserById(payload.userId);
+    if (!user) {
+      throw new BadRequestException('Invalid accessToken');
+    }
+
     const role = await this.rolesService.requestRole(user.id, arrowId)
     const arrow = await this.arrowsService.getArrowById(arrowId);
 
@@ -102,13 +119,20 @@ export class RolesResolver {
     return role;
   }
 
-  @UseGuards(GqlAuthGuard)
   @Mutation(() => Role, {name: 'removeRole'})
   async removeRole(
-    @CurrentUser() user: User,
+    @Args('accessToken') accessToken: string,
     @Args('sessionId') sessionId: string,
     @Args('roleId') roleId: string,
   ) {
+    const payload = this.jwtService.verify(accessToken, {
+      secret: this.configService.get('JWT_ACCESS_TOKEN_SECRET')
+    });
+    const user = await this.usersService.getUserById(payload.userId);
+    if (!user) {
+      throw new BadRequestException('Invalid accessToken');
+    }
+    
     const role = await this.rolesService.removeRole(user.id, roleId)
     const removedUser = await this.usersService.getUserById(role.userId);
     const arrow = await this.arrowsService.getArrowById(role.arrowId);
