+++ src/tabs/tabs.resolver.ts
@@ -1,17 +1,21 @@
 import { Args, Int, Mutation, Parent, ResolveField, Resolver } from '@nestjs/graphql';
 import { Arrow } from 'src/arrows/arrow.model';
 import { ArrowsService } from 'src/arrows/arrows.service';
-import { CurrentUser, GqlAuthGuard } from 'src/auth/gql-auth.guard';
 import { Tab } from './tab.model';
 import { TabsService } from './tabs.service';
-import { User as UserEntity } from '../users/user.entity';
-import { BadRequestException, UseGuards } from '@nestjs/common';
+import { BadRequestException } from '@nestjs/common';
 import { RemoveTabResult } from './dto/remove-tab-result.dto';
 import { SheafsService } from 'src/sheafs/sheafs.service';
+import { JwtService } from '@nestjs/jwt';
+import { ConfigService } from '@nestjs/config';
+import { UsersService } from 'src/users/users.service';
 
 @Resolver(Tab)
 export class TabsResolver {
   constructor(
+    private readonly configService: ConfigService,
+    private readonly jwtService: JwtService,
+    private readonly usersService: UsersService,
     private readonly tabsService: TabsService,
     private readonly arrowsService: ArrowsService,
     private readonly sheafsService: SheafsService,
@@ -24,14 +28,21 @@ export class TabsResolver {
     return this.arrowsService.getArrowById(tab.arrowId);
   }
 
-  @UseGuards(GqlAuthGuard)
   @Mutation(() => [Tab], {name: 'createGraphTab'})
   async createGraphTab(
-    @CurrentUser() user: UserEntity,
+    @Args('accessToken') accessToken: string,
     @Args('name') name: string,
     @Args('routeName') routeName: string,
     @Args('arrowId', {nullable: true}) arrowId: string,
   ) {
+    const payload = this.jwtService.verify(accessToken, {
+      secret: this.configService.get('JWT_ACCESS_TOKEN_SECRET')
+    });
+    const user = await this.usersService.getUserById(payload.userId);
+    if (!user) {
+      throw new BadRequestException('Invalid accessToken');
+    }
+
     if (arrowId) {
       let arrow = await this.arrowsService.getArrowById(arrowId);
       if (!arrow) {
@@ -63,15 +74,23 @@ export class TabsResolver {
     }
 
   }
-  @UseGuards(GqlAuthGuard)
+
   @Mutation(() => [Tab], { name: 'createTab' })
   async createTab(
-    @CurrentUser() user: UserEntity,
+    @Args('accessToken') accessToken: string,
     @Args('arrowId') arrowId: string,
     @Args('i', {type: () => Int, nullable: true}) i: number,
     @Args('isFrame') isFrame: boolean,
     @Args('isFocus') isFocus: boolean,
   ) {
+    const payload = this.jwtService.verify(accessToken, {
+      secret: this.configService.get('JWT_ACCESS_TOKEN_SECRET')
+    });
+    const user = await this.usersService.getUserById(payload.userId);
+    if (!user) {
+      throw new BadRequestException('Invalid accessToken');
+    }
+
     const arrow = await this.arrowsService.getArrowById(arrowId);
     if (!arrow) {
       throw new BadRequestException('Arrow not found');
@@ -82,15 +101,22 @@ export class TabsResolver {
     return this.tabsService.appendTab(user, arrow, isFrame, isFocus);
   }
 
-  @UseGuards(GqlAuthGuard)
   @Mutation(() => [Tab], { name: 'createTabByRouteName' })
   async createTabByRouteName(
-    @CurrentUser() user: UserEntity,
+    @Args('accessToken') accessToken: string,
     @Args('routeName') routeName: string,
     @Args('i', {type: () => Int, nullable: true}) i: number,
     @Args('isFrame') isFrame: boolean,
     @Args('isFocus') isFocus: boolean,
   ) {
+    const payload = this.jwtService.verify(accessToken, {
+      secret: this.configService.get('JWT_ACCESS_TOKEN_SECRET')
+    });
+    const user = await this.usersService.getUserById(payload.userId);
+    if (!user) {
+      throw new BadRequestException('Invalid accessToken');
+    }
+
     const arrow = await this.arrowsService.getArrowByRouteName(routeName);
     if (!arrow) {
       throw new BadRequestException('Arrow not found');
@@ -101,24 +127,36 @@ export class TabsResolver {
     return this.tabsService.appendTab(user, arrow, isFrame, isFocus);
   }
 
-  @UseGuards(GqlAuthGuard)
   @Mutation(() => [Tab], { name: 'updateTab' })
   async updateTab(
-    @CurrentUser() user: UserEntity,
+    @Args('accessToken') accessToken: string,
     @Args('tabId') tabId: string,
     @Args('i', {type: () => Int}) i: number,
     @Args('isFrame') isFrame: boolean,
     @Args('isFocus') isFocus: boolean,
   ) {
+    const payload = this.jwtService.verify(accessToken, {
+      secret: this.configService.get('JWT_ACCESS_TOKEN_SECRET')
+    });
+    const user = await this.usersService.getUserById(payload.userId);
+    if (!user) {
+      throw new BadRequestException('Invalid accessToken');
+    }
     return this.tabsService.updateTab(user, tabId, i, isFrame, isFocus);
   }
 
-  @UseGuards(GqlAuthGuard)
   @Mutation(() => RemoveTabResult, { name: 'removeTab' })
   async removeTab(
-    @CurrentUser() user: UserEntity,
+    @Args('accessToken') accessToken: string,
     @Args('tabId') tabId: string,
   ) {
+    const payload = this.jwtService.verify(accessToken, {
+      secret: this.configService.get('JWT_ACCESS_TOKEN_SECRET')
+    });
+    const user = await this.usersService.getUserById(payload.userId);
+    if (!user) {
+      throw new BadRequestException('Invalid accessToken');
+    }
     return this.tabsService.removeTab(user, tabId);
   }
 }
